+++ f1t/settings/development.py
@@ -1,6 +1,6 @@
 import mimetypes
 
-from .base import *
+from ._base import *
 
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = "INSECURE_DJANGO_SECRET_KEY_TO_BE_USED_AT_LOCAL_ENV"