+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - __core__ directory to f1t
 - static and templates directories to project folder
+### Fixed
+- 403.html
 
 ## [1.8.4] - 2023-08-03
 ### Added